RAG_vs_SQL/requirements.txt
langchain>=0.0.267
langchain-community>=0.0.13
langchain-openai>=0.0.2
openai>=1.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
chromadb>=0.4.18
langchain-experimental>=0.0.37
pandas>=2.0.0
sqlalchemy>=2.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1 